Fix missing text
The issue was that we use mem::take to take the text rect requests,
which set the scissors to the default value of 0.
The next time, when we chunk by the scissor all consumed groups will be
merged into a single group. So when we use skip it's gonna skip more
than intended, if there were at least 2 groups.